refactor: remove legacy artifacts now that the new parser is the only parser
Several leftovers from the parser-swap transition were still in the
tree:

- parser_new_test.go held a 'new vs legacy' parity test. Both code
  paths run through the same parser now, so the test was a tautology.
  Removed.

- parser_new.go was a misleading filename — it's the *only* parser.
  Merged its contents into parser.go.

- parser.go was a mix of AST type definitions and parser entry points.
  Split: AST types move to ast.go (84 lines), parser orchestrator stays
  in parser.go (488 lines).

- Comments throughout the package referenced 'the legacy parser' or
  'legacy parity'. Updated to describe current behavior on its own
  terms.

- A redundant parseNodes wrapper around parseNodesUntil was dropped
  (only one call site, inlined).

- Restored missing doc comment on IndentConfig, updated doc.go's
  architecture description (no more 'parser_new.go', no more 'legacy
  combined file').

No behavior change: all 45 testdata fixtures still byte-identical, the
shopware/storefront smoke test still 0/0/0 on 326 files, all 27 caller
files in internal/verifier/ unchanged.
