+++ internal/html/ast.go
@@ -1,0 +1,84 @@
+// AST node types for the html package. Each node has a Dump(int) string
+// method (defined in format.go) that renders it back to its textual form.
+package html
+
+// Attribute represents an HTML attribute with key and value.
+type Attribute struct {
+	Key   string
+	Value string
+}
+
+// Node is the interface for nodes in the AST.
+type Node interface {
+	Dump(indent int) string
+}
+
+// NodeList is a sequence of AST nodes. It has its own Dump method that
+// arranges children with appropriate inter-node whitespace.
+type NodeList []Node
+
+// ConfiguredNodeList wraps a NodeList with the IndentConfig under which it
+// should format. NewAdminParser and NewStorefrontParser return this so the
+// caller can Dump(0) without managing config separately.
+type ConfiguredNodeList struct {
+	Nodes  NodeList
+	Config IndentConfig
+}
+
+// RawNode holds verbatim source text — anything outside structured HTML or
+// Twig syntax, plus the bytes of any Twig tag that has no registered handler.
+type RawNode struct {
+	Text string
+	Line int
+}
+
+// CommentNode represents an HTML <!-- ... --> comment.
+type CommentNode struct {
+	Text string
+	Line int
+}
+
+// TemplateExpressionNode represents a `{{ ... }}` Twig expression.
+type TemplateExpressionNode struct {
+	Expression string
+	Line       int
+}
+
+// ElementNode represents an HTML element.
+type ElementNode struct {
+	Tag         string
+	Attributes  NodeList
+	Children    NodeList
+	SelfClosing bool
+	// Unclosed reports that the element opened with `<tag>` but its
+	// children parser yielded on an outer Twig terminator (e.g.
+	// `{% endblock %}`) before reaching `</tag>`. The closing tag is
+	// elsewhere in the source, typically wrapped in another control-flow
+	// block. The formatter therefore does NOT emit `</tag>` for unclosed
+	// elements — the matching `</tag>` lives as a RawNode further down.
+	Unclosed bool
+	Line     int
+}
+
+// TwigBlockNode represents `{% block name %}...{% endblock %}`.
+type TwigBlockNode struct {
+	Name     string
+	Children NodeList
+	Line     int
+}
+
+// TwigIfNode represents `{% if %}...{% elseif %}...{% else %}...{% endif %}`.
+// Branches are stored as parallel slices ElseIfConditions / ElseIfChildren.
+type TwigIfNode struct {
+	Condition        string
+	Children         NodeList
+	ElseIfConditions []string
+	ElseIfChildren   []NodeList
+	ElseChildren     NodeList
+	Line             int
+}
+
+// ParentNode represents `{% parent %}` or `{% parent() %}`.
+type ParentNode struct {
+	Line int
+}
